Update ShopStatsCalculator
Calculates the sum, average, variance and standard deviation of sales made per day. Also calculates the sum, average, and variance of the revenue made per sale (per day or per month, up to the user input) Calculates if the total profit made in that interval met the goal profit value.